Backend/requirements.txt
google-api-python-client>=1.7.9